Refactor app startup and authentication flow
Moved app initialization and routing logic from main.dart into a new MyApp widget in app.dart. Introduced AuthCubit and AuthState for managing authentication state using Bloc, and added AppWrapper to handle navigation based on authentication status. Added AuthService.isAuthenticated() for checking login state. This refactor improves separation of concerns and centralizes authentication logic.
