+++ lib/main.dart
@@ -1,29 +1,16 @@
-import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:hive_flutter/adapters.dart';
 import 'package:isar/isar.dart';
+import 'package:yack/providers/auth/auth_cubit.dart';
 import 'package:yack/providers/auth/confirm_cubit.dart';
 import 'package:yack/providers/auth/login_cubit.dart';
 import 'package:yack/providers/auth/password_reset_cubit.dart';
 import 'package:yack/providers/auth/signup_cubit.dart';
-import 'package:yack/screens/auth/confirm.dart';
 import 'package:firebase_core/firebase_core.dart';
-import 'package:yack/screens/contract_agr/contract_agreement.dart';
-import 'package:yack/screens/contract_agr/nomore_contracts.dart';
-import 'package:yack/screens/root.dart';
+import 'app.dart';
 import 'firebase_options.dart';
-import 'package:yack/screens/auth/forgetPassword.dart';
-import 'package:yack/screens/auth/login.dart';
-import 'package:yack/screens/auth/signup.dart';
-import 'package:yack/screens/welcome.dart';
-import '../theme/theme.dart';
-import 'screens/sign_contract.dart';
-import 'screens/settings.dart';
-import 'screens/create_contract.dart';
-import 'screens/scan_contract.dart';
 import 'utils/translation_handler.dart';
 import 'package:yack/db/models/contract.dart';
 import 'package:yack/db/models/mediaFile.dart';
@@ -70,34 +57,18 @@ void main() async {
 
   // Initialize Hive
   await Hive.initFlutter();
+
   final userBox = await Hive.openBox('user');
 
+
+  // Initialize Translation Handler
   await TranslationHandler.initialize(userBox);
-
-  late final String initialRoute;
-
-  final user = FirebaseAuth.instance.currentUser;
-
-  if (userBox.get('didFirstTime') == null || userBox.get('didFirstTime') == false) {
-    // First time opening the app → show welcome
-    initialRoute = '/welcome';
-    await userBox.put('didFirstTime', true);
-  } else if (user != null) {
-    // User already logged in → go directly to home
-    initialRoute = '/home';
-  } else if (userBox.get('didFirstLogin') == true) {
-    // User has seen login before → go to login
-    initialRoute = '/login';
-  } else {
-    // Otherwise → signup
-    initialRoute = '/signup';
-  }
 
 
   runApp(
       MultiBlocProvider(
           providers: [
-              // BlocProvider<AuthCubit>(create: (_) => AuthCubit()),
+              BlocProvider<AuthCubit>(create: (_) => AuthCubit()..checkAuth()),
               BlocProvider<LoginCubit>(create: (_) => LoginCubit()),
               BlocProvider<SignupCubit>(create: (_) => SignupCubit()),
               BlocProvider<ConfirmCubit>(create: (_) => ConfirmCubit(),),
@@ -107,119 +78,8 @@ void main() async {
               // BlocProvider<ContractCubit>(create: (_) => ContractCubit()),
           // add others here...,
           ],
-          child: MyApp(initialRoute: initialRoute)
+          child: MyApp(userBox: userBox,)
       )
   );
-}
-
-class MyApp extends StatefulWidget {
-  final String initialRoute;
-  const MyApp({super.key, required this.initialRoute});
-
-  @override
-  State<MyApp> createState() => _MyAppState();
-}
-
-class _MyAppState extends State<MyApp> {
-  late Box userBox;
-
-  @override
-  void initState() {
-    super.initState();
-    userBox = Hive.box('user');
-  }
-
-  ThemeMode get _themeMode {
-    final themeValue = userBox.get('theme');
-    switch (themeValue) {
-      case 1:
-        return ThemeMode.light;
-      case 2:
-        return ThemeMode.dark;
-      case 3:
-      case null:
-      default:
-        return ThemeMode.system;
-    }
-  }
-
-  /// Wraps each screen with a theme-aware system UI style
-  Widget themedRoute(BuildContext context, Widget child, {
-   bool transparent = false
-  }) {
-    final brightness = MediaQuery.of(context).platformBrightness;
-    final isDarkMode = _themeMode == ThemeMode.dark ||
-        (_themeMode == ThemeMode.system && brightness == Brightness.dark);
-
-    final overlayStyle = SystemUiOverlayStyle(
-      statusBarColor: Colors.transparent,
-      systemNavigationBarColor: !transparent? Theme.of(context).colorScheme.surface: Colors.transparent,
-      systemNavigationBarDividerColor: !transparent?Theme.of(context).colorScheme.surface: Colors.transparent,
-      systemNavigationBarContrastEnforced: false,
-      statusBarIconBrightness: isDarkMode ? Brightness.light : Brightness.dark,
-      systemNavigationBarIconBrightness:
-      isDarkMode ? Brightness.light : Brightness.dark,
-      statusBarBrightness: isDarkMode ? Brightness.dark : Brightness.light,
-    );
-
-    return AnnotatedRegion<SystemUiOverlayStyle>(
-      value: overlayStyle,
-      child: child,
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    // ValueListenableBuilder updates MaterialApp when the Hive value changes
-    return ValueListenableBuilder(
-      valueListenable: userBox.listenable(keys: ['theme','language']),
-      builder: (context, box, _) {
-            return MaterialApp(
-              debugShowCheckedModeBanner: false,
-              theme: AppTheme.lightTheme,
-              darkTheme: AppTheme.darkTheme,
-              themeMode: _themeMode,
-              locale: TranslationHandler.locale,
-              supportedLocales: TranslationHandler.supportedLocales.toList(),
-              localizationsDelegates: const [
-                GlobalMaterialLocalizations.delegate,
-                GlobalCupertinoLocalizations.delegate,
-                GlobalWidgetsLocalizations.delegate,
-              ],
-              routes: {
-            '/welcome': (context) =>
-                themedRoute(context, const OnboardingScreen(),transparent: true),
-
-            // AUTH ROUTES
-            '/signup': (context) =>
-                themedRoute(context,  SignUpScreen(),transparent: true),
-            '/login': (context) =>
-                themedRoute(context,  LoginScreen(),transparent: true),
-            '/confirm': (context) =>
-                themedRoute(context,  ConfirmAccount(),transparent: true),
-            '/forgot-password': (context) =>
-                themedRoute(context,  ForgetPassword(),transparent: true),
-
-            // CONTRACT ROUTES
-            '/contract/scan_contract': (context) =>
-                themedRoute(context, const ScanContractScreen()),
-            '/contract/sign_contract': (context) =>
-                themedRoute(context, const SignContractScreen()),
-            '/contract/create_contract': (context) =>
-                themedRoute(context, const CreateContractScreen()),
-            '/contract/view': (context) => const ContractAgreement(),
-
-            '/settings': (context) =>
-                themedRoute(context, const SettingsScreen()),
-            '/upgrade': (context) =>
-                themedRoute(context, const NoMoreContractsAvailable(),transparent: true),
-            '/home': (context) =>
-                themedRoute(context, const BottomNavBar()),
-          },
-          initialRoute: widget.initialRoute,
-            );
-          },
-    );
-  }
 }
