+++ lib/providers/auth/auth_state.dart
@@ -1,0 +1,22 @@
+import 'package:equatable/equatable.dart';
+
+abstract class AuthState extends Equatable {
+  @override
+  List<Object?> get props => [];
+}
+
+class AuthInitial extends AuthState {}
+
+class AuthChecking extends AuthState {}
+
+class Authenticated extends AuthState {}
+
+class Unauthenticated extends AuthState {}
+
+class AuthError extends AuthState {
+  final String messageKey;
+  AuthError(this.messageKey);
+
+  @override
+  List<Object?> get props => [messageKey];
+}
